face_verification/requirements.txt
face_recognition==1.3.0
facenet_pytorch==2.5.2
numpy==1.20.1
opencv_python==4.6.0.66
Pillow==9.2.0
torch==1.12.1
torchvision==0.13.1
